% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hi_recreation_init_conda_helper.R
\name{hi_recreation_init_conda_helper}
\alias{hi_recreation_init_conda_helper}
\title{Conda environment helper}
\usage{
hi_recreation_init_conda_helper(env_name = "wild_hi_recreation")
}
\arguments{
\item{env_name}{string; the name you want to use for the environment.}
}
\value{
This function returns the environment name.
}
\description{
The function, `hi_recreation_init_conda_helper()`, creates the appropriate Python environment for the user in miniconda. If the user does not have miniconda installed, it handles this too.
}
\details{
Using `hi_recreation_init_conda_helper()` may require administrator privileges. It creates a conda environment with a specified environment name using an environment.yml file that is included with the package. If miniconda is not downloaded, it will install miniconda. This can take some time.
}
\examples{
\dontrun{
envName <- hi_recreation_init_conda_helper(env_name = 'hi_rec')
}

}
