% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildlifeHI-package.r
\docType{package}
\name{wildlifeHI-package}
\alias{wildlifeHI-package}
\title{wildlifeHI: Tools for Linking Wildlife Tracking Data to Human Infrastructure}
\description{
This package contains a suite of basic tools for studying how wildlife are impacted by human infrastructure using tracking data. It focuses largely on working with Open Street Map (OSM) data using the package \code{osmdata}.
}
\details{
The package \code{wildlifeHI} incorporates a basic wrapper function (\code{hi_get_osm}) for extracting OSM data using the \code{osmdata} package. It also has functions for performing different types of analysis to study how wildlife respond to human infrastructure; these include distance analysis, identifying and mapping crossings, and buffer-type analysis. The goal is to include several of these functions as MoveApps.
}
\author{
Jed Long
}
