% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hi_check_bbox.R
\name{hi_check_bbox}
\alias{hi_check_bbox}
\title{Compare Bounding Boxes by trackId}
\usage{
hi_check_bbox(move)
}
\arguments{
\item{move}{an object of the class \code{move}. For more information on objects of this type see \code{help(move)}.}
}
\value{
This function returns a \code{sf} POLYGON object containing the bounding boxes for each individual \code{trackId}within the tracking data.
}
\description{
This function is a a checking tool to investigate the bounding boxes of each \code{trackId} within a \code{MoveStack}.
}
\details{
This is a simple tool to look at the bounding boxes of each individual within a \code{MoveStack}. This is useful to explore how large different bounding boxes are prior to performing queries for OSM data using \code{hi_get_osm}. This is often useful when the overall study area is large, but the bounding boxes of individual trajectories are relatively small in comparison.
}
\examples{
library(move)
data(fishers)
fishers_bb <- hi_check_bbox(fishers) 

\dontrun{
library(mapview)
mapview(fishers_bb['trackId'])
}


}
