% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hi_recreation.R
\name{hi_recreation}
\alias{hi_recreation}
\title{HI Recreation}
\usage{
hi_recreation(move, cookies = NULL, env_name = "wild_hi_recreation")
}
\arguments{
\item{move}{an object of the class \code{move2}. For more information on objects of this type see \code{help(move2)}.}

\item{cookies}{string; formatted string from `hi_recreation_cookie_constructor()`. This is a required parameter.}

\item{env_name}{string (optional); the name of the environment specified in `hi_recreation_init_conda_helper()` or the name of a Conda environment that meets the requirements.}
}
\value{
This function returns a \code{move2} object with the R, G, B, and A pixel values and the calculated intensity values appended from the Strava Global Heatmap.
}
\description{
The main function, `hi_recreation()`, calculates the intensity of outdoor recreation along the GPS track provided.
}
\details{
To use the `hi_recreation()` function, you will need access to a Strava account that has had some activity.
 In `hi_recreation()`, pixel values are obtained from the Strava Global Heatmap and an intensity of use is calculated based on those pixel values along the GPS track provided as input. This can be useful for determining if individuals avoid or prefer areas of high human activity.
}
\examples{
\dontrun{
library(move)
library(move2)
data(fishers)
fishers2 <- mt_as_move2(fishers)
hi_recreation_init_conda_helper(env_name = 'hi_rec')
cookies <- hi_recreation_cookie_constructor(<Key-Pair-Id>, <Policy>, <Signature>)
fishers2 <- hi_recreation(move = fishers2, cookies = cookies, env_name = 'hi_rec')
}

}
